[#179] feat 코드 정렬
